+++ packages/react-server/examples/basic/src/routes/test/dynamic/_cilent.tsx
@@ -1,0 +1,8 @@
+"use client";
+
+import { useRouter } from "@hiogawa/react-server/client";
+
+export function ClientLocation() {
+  const location = useRouter((s) => s.location);
+  return <>{location.pathname}</>;
+}