+++ packages/react-server/examples/basic/src/routes/test/dynamic/_utils.tsx
@@ -1,4 +1,5 @@
 import type { PageProps } from "@hiogawa/react-server/server";
+import { ClientLocation } from "./_cilent";
 
 export function TestDynamic({
   file: file,
@@ -10,7 +11,13 @@ export function TestDynamic({
   return (
     <div className="flex flex-col gap-1">
       <div>file: {file}</div>
-      <div>pathname: {new URL(props.request.url).pathname}</div>
+      <div>
+        pathname:{" "}
+        {props.request.url.slice(new URL(props.request.url).origin.length)}
+      </div>
+      <div>
+        pathname (client): <ClientLocation />
+      </div>
       <div>params: {JSON.stringify(props.params)}</div>
     </div>
   );