+++ packages/react-server/examples/basic/src/routes/test/dynamic/layout.tsx
@@ -11,8 +11,14 @@ export default function Layout(props: LayoutProps) {
           "/test/dynamic",
           "/test/dynamic/static",
           "/test/dynamic/abc",
+          // these two should work same inside the application?
+          "/test/dynamic/✅",
+          "/test/dynamic/" + encodeURI("✅"),
           "/test/dynamic/abc/def",
         ]}
+        activeProps={{
+          "aria-current": "page",
+        }}
       />
       <div>{props.children}</div>
     </div>